# Fix pool elements leftover info bug