@SyntaxNode 2020-05-26
The change looks pretty conclusive.. it's just that according to the Go doc's, it shouldn't be needed:

```
To unmarshal a JSON array into a slice, Unmarshal resets the slice length to zero and then appends each element to the slice. As a special case, to unmarshal an empty JSON array into a slice, Unmarshal replaces the slice with a new empty slice.
```

I'm a bit scared we're not seeing that behavior.