@hhhjort 2020-05-26
Oh, I see why we would need this change even if the documented behavior is followed exactly. We start with a request that has a non-empty `puts`. The pool object gets recycled at the end, and then gets reused to service the request `{}`. When it tries unmarshalling that, it doesn't find the `puts` object, so it won't try overwriting the `Puts` field with an empty slice. The result is an unmarshalled put that has items in the `Puts` slice. A request that looks like `{"puts":[]}` should always be empty given the documented behavior. I would be curious if the old code would fail the test if that was the only "empty" request.