@guscarreon 2020-05-26
>  I would be curious if the old code would fail the test if that was the only "empty" request.
The "old code" does not fail the tests that this PR adds.