@SyntaxNode 2020-05-27
@hhhjort Ah, ok. I'm not sure if that's the same problem we saw in production. I have a hard time believing we were sending a large amount of empty requests. That being said, there might've been some strange race condition with the slice which is resolved by this.

This looks good to me. I think we might still want to test for race conditions or continue investigating to prove it's no longer a concern.